Refactored classes and removed garbage
